Much enhanced rope blocks, added rope ladder
I did a bunch of work enhancing the rope blocks in this. I gave them a
nodebox model, limited the length the rope will extend to, allowed
blocks with rope of varying length to be crafted, and fixed a couple of
problems with ropes - namely that removing the block wouldn't
necessarily remove the whole rope (now there's a "rope top" that eats
its way down to ensure the whole rope is removed) and that a rope
couldn't be cut once it was in place. I also added a rope ladder
variant, and made it possible to craft ropes using cotton as well as
vines.

I may have broken the vines part of this mod in the process, I'll do
further testing and code cleanup soon.